Rewrite argument parsing in order to more easily set desired gold amount.
Argument parsing is rewritted using native argparse module instead of
native and previous getopt module.
It permits to add more easily new arguments as gold amount or an item name.
Parsing is done directly in __main__, which is rather compact. Previous
start() function, which was dedicated to this parsing, is removed.
Arguments validation is maintained.

New commands are added in order to get or set gold amount, and to list a
given item/collectable/etc.

All functions use the savefile first argument instead of global file.