Adding debug mode
The -d or --debug argument enters in debug mode where the global variable debug can be used in every function.
First application appears in checkItemName() function where, when debug mode is on, the item category: Collectable, Material or KeyItem, is shown as well as its index (from the dictionary in itemlist.py).