Factorizing item categories
There are similar parts of code in order to treat item categories (Collectable, Material, KeyItem) separately. Most of these parts can be melt in one code within a loop on every categories, if content of these categories are factorized in a common AllItems new dictionary.
This AllItems dictionary also provides, for each category, some values specific to a given category, e.g., the maximum number of slots available in the monado01 backup file, or where to start searching the first item of the category in this file.
Using AllItems is first applied in the checkItemName() function.