Validation of new number of items asked while setting them
The AllItems dictionary has a new parameter : maxNb, which sets the maximum number of allowed items for each category,
i.e., 99 for Collectable and Material, 1 for KeyItem.
This permits also to add new items from KeyItem category as while not breaking anything.

The setItem function is able to add new items if the length of new inventory is not too long.

Another validation is added to prevent choice 0 for number of items asked, as removing an existing item is not possible yet.