Factorizing the writing to the savefile
The new function writeItems(savefile,myItems) permits to factorize every code block in colony6, setItem and deleteItem functions where, when the dictionary of all categorized items with their respective number value is set, these items and values are transformed into a binary string to be written to the backup file monado01,02 or 03.

Also, we add validation on deleting an item, i.e., we check that this item really exists in the inventory before deleting it. In case of trouble, a message is shown to the user.